requirements.txt
pandas
numpy
requests
feedparser
scikit-learn
joblib
streamlit
python-dateutil
